refactor(companion): extract shared browser-instance library
Extract the Chrome instance management duplicated across mermaid,
diagrams.net, bpmn and excalidraw into lib/browser-instance, consumed
as a file: dependency. The library has zero dependencies on purpose:
npm installs file: packages as symlinks, so imports resolve from the
package's real location, outside the service node_modules — each
service injects its own puppeteer module and logger instead.

Each service keeps a thin src/browser-instance.js shim that declares
its divergence explicitly (env prefix, extra Chrome flags such as
--disable-web-security for diagrams.net and bpmn).

This also ports the browser leak fix to bpmn and excalidraw, which
were still launching Chrome once at startup with no recovery when the
process died: they now use the memoized launch promise, relaunch on
demand, bound CDP calls with protocolTimeout and route the Chrome
process output to the logger.

Docker builds resolve the library through a named "lib" build context
in docker-bake.hcl, copied to /usr/local/lib/browser-instance so the
file:../lib/browser-instance dependency resolves identically locally
and in the image. Note that the companion images now require buildx
bake (a plain docker build lacks the named context).
