+++ bpmn/package.json
@@ -19,6 +19,7 @@
     "url": "https://github.com/yuzutech/kroki.git"
   },
   "dependencies": {
+    "@kroki/browser-instance": "file:../lib/browser-instance",
     "bpmn-js": "18.18.0",
     "micro": "10.0.1",
     "pino": "9.14.0",
