+++ diagrams.net/package.json
@@ -17,6 +17,7 @@
     "url": "https://github.com/yuzutech/kroki.git"
   },
   "dependencies": {
+    "@kroki/browser-instance": "file:../lib/browser-instance",
     "micro": "10.0.1",
     "pino": "9.14.0",
     "pino-debug": "^2.0.0",
