+++ excalidraw/src/worker.js
@@ -1,12 +1,12 @@
 import { URL, fileURLToPath } from 'node:url'
 import puppeteer from 'puppeteer'
 import { logger } from './logger.js'
+import { getBrowserWSEndpoint, protocolTimeout } from './browser-instance.js'
 
 const __dirname = fileURLToPath(new URL('.', import.meta.url))
 
 export default class Worker {
-  constructor(browserInstance) {
-    this.browserWSEndpoint = browserInstance.wsEndpoint()
+  constructor() {
     this.pageUrl =
       process.env.KROKI_EXCALIDRAW_PAGE_URL || 'http://127.0.0.1:8004/public/index.html'
     this.assetPath = process.env.KROKI_EXCALIDRAW_ASSET_PATH || '/public'
@@ -14,10 +14,10 @@ export default class Worker {
 
   async convert(task) {
     const browser = await puppeteer.connect({
-      browserWSEndpoint: this.browserWSEndpoint,
+      browserWSEndpoint: await getBrowserWSEndpoint(),
       ignoreHTTPSErrors: true,
       // Bound CDP calls so a wedged Chrome fails fast instead of stalling 180s.
-      protocolTimeout: Number(process.env.KROKI_EXCALIDRAW_PROTOCOL_TIMEOUT) || 30000
+      protocolTimeout
     })
     const page = await browser.newPage()
     try {
