+++ mermaid/package.json
@@ -19,14 +19,15 @@
     "url": "https://github.com/yuzutech/kroki.git"
   },
   "dependencies": {
+    "@kroki/browser-instance": "file:../lib/browser-instance",
     "micro": "10.0.1",
     "pino": "9.14.0",
     "pino-debug": "2.0.0",
     "puppeteer": "25.1.0"
   },
   "devDependencies": {
-    "mermaid": "11.15.0",
     "@biomejs/biome": "2.4.16",
+    "mermaid": "11.15.0",
     "pngjs": "7.0.0"
   },
   "volta": {
